Massive update for Lily 1.0
Lily has gone through a lot of breaking api changes on the march to
version 1.0. Most are internal, but a few are minor and syntax
related. Now that the breaking is (hopefully) done, this pulls these
libraries up to the latest Lily.

Most of these scripts work. `lapp.lily` is a casualty: I don't know
enough of it at a quick glance to see why it's failing. The rest I've
done baseline testing against and they appear to work.